Added new vg-crossorigin directive to let dev set a cross origin policy for any media file or track. close #123
Optional directive for `vg-media` to add or remove a crossorigin policy to the video object. Possible values are: "anonymous" and "use-credentials". This feature should be enabled if you want to have your subtitles or video files on a different domain than the video player. Additionally you need to add CORS policies to your video and track files to your server to make it work.

I recommend this Chrome extension for testing purposes https://chrome.google.com/webstore/detail/allow-control-allow-origi/nlfbmbojpeacfghkpbjhddihlkkiljbi/reviews

This was a hard one... Really.