Update vg-controller.js
The autoplay watcher $scope.$watch("vgAutoPlay", this.onUpdateAutoPlay.bind(this)); immediately runs onUpdateAutoPlay which runs play, while the mediaElement is not yet initialized by the vgMedia directive. Resulting in an error stating: "mediaElement[0] undefined".  

This commit adds a mediaElement[0] stub which is overwritten when the real mediaElement is initialized. (This also solves the clearMedia undefined error).