feat(vgMedia): vgNativePlayBlacklist configuration to supply an array of test functions for blacklis
You may now supply vgNativePlayBlacklist on vgMedia to prevent native playback of sources as
desired. It expects an array of functions accepting  the media source and user agent. If any of
these return true,  native playback will not be attempted. This  may be useful in combination with
playback plugins. For example, you may want to use hls.js  in Android Chrome rather than its native
handling.